Struct memfs::MemFile

pub struct MemFile { /* private fields */ }

The struct that represents a file in memory that is backed by MappedPages

Implementations

impl MemFile

pub fn create(name: String, parent: &DirRef) -> Result<FileRef, &'static str>

Allocates writable memory space for the given contents and creates a new file containing that content in the given parent directory.

pub fn from_mapped_pages( mapped_pages: MappedPages, name: String, len: usize, parent: &DirRef ) -> Result<FileRef, &'static str>

Creates a new MemFile in the given parent directory with the contents of the given mapped_pages.